Hacked it up to start tearing out SharpGL which has major issues with WPF.
To use SharpGL with WPF, you have to either use a FBO or a Blitted device. The Blitted device has fine performance but rolls back to OpenGL 1.1, thus no extensions (like glGenBuffers).

The FBO has the glGenBuffers extension, but it also introduces a 3-5 second lag after closing the application and waiting on the debugger. Rather annoying. Going back to OpenTK with a WindowsFormHost. Hopefully the WFH won't present an issue later on. It's a shame we're losing the OO-ness of SharpGL, but all SharpGL was doing on the backend was calling Context::MakeCurrent() anyways before each OpenGL instance usage, so you can do the same thing (albeit manually) with OpenTK to get multi-contexts.