Fix Culture Problems
By not specifying the Culture, a lot of formatting and parsing is
completely different in certain cultures, causing a lot of things to
break. The Shader Generator generates wrong shaders in Germany for
example, as we Germans use a "," as the decimal point, instead of a ".".
This obviously causes a lot of problems with the floating point numbers,
as the shader compiler now tried to parse them as twice as many integer
parameters. Other issues like Decimal Numbers in the UI caused problems
too, so the safest solution is to just set the Culture globally to an
invariant Culture.